+++ engine/flink/management/periodic/src/main/scala/pl/touk/nussknacker/engine/management/periodic/DeploymentActor.scala
@@ -2,6 +2,7 @@ package pl.touk.nussknacker.engine.management.periodic
 
 import akka.actor.{Actor, Props}
 import com.typesafe.scalalogging.LazyLogging
+import pl.touk.nussknacker.engine.management.periodic.db.ScheduledRunDetails
 import pl.touk.nussknacker.engine.management.periodic.DeploymentActor.{CheckToBeDeployed, DeploymentCompleted, WaitingForDeployment}
 
 import scala.concurrent.duration._
@@ -14,7 +15,7 @@ object DeploymentActor {
 
   case object CheckToBeDeployed
 
-  case class WaitingForDeployment(ids: List[PeriodicProcessDeploymentId])
+  case class WaitingForDeployment(ids: List[ScheduledRunDetails])
 
   case class DeploymentCompleted(success: Boolean)
 }
@@ -31,18 +32,18 @@ class DeploymentActor(service: PeriodicProcessService, interval: FiniteDuration)
   override def receive: Receive = {
     case CheckToBeDeployed =>
       logger.debug("Checking processes to be deployed")
-      service.findToBeDeployed.foreach { ids => self ! WaitingForDeployment(ids.toList) }
+      service.findToBeDeployed.foreach { runDetails => self ! WaitingForDeployment(runDetails.toList) }
     case WaitingForDeployment(Nil) =>
-    case WaitingForDeployment(deploymentId :: _) =>
-      logger.info("Found a process to be deployed: {}", deploymentId)
-      service.deploy(deploymentId) onComplete { result => self ! DeploymentCompleted(result.isSuccess) }
-      context.become(ongoingDeployment(deploymentId))
+    case WaitingForDeployment(runDetails :: _) =>
+      logger.info("Found a process to be deployed: {}", runDetails.processDeploymentId)
+      service.deploy(runDetails) onComplete { result => self ! DeploymentCompleted(result.isSuccess) }
+      context.become(ongoingDeployment(runDetails))
   }
 
-  private def ongoingDeployment(id: PeriodicProcessDeploymentId): Receive = {
+  private def ongoingDeployment(runDetails: ScheduledRunDetails): Receive = {
     case CheckToBeDeployed =>
     case DeploymentCompleted(success) =>
-      logger.info("Deployment {} completed, success: {}", id, success)
+      logger.info("Deployment {} completed, success: {}", runDetails.processDeploymentId, success)
       context.unbecome()
   }
 }