+++ engine/flink/management/periodic/src/main/scala/pl/touk/nussknacker/engine/management/periodic/PeriodicProcessManager.scala
@@ -12,6 +12,7 @@ import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus
 import pl.touk.nussknacker.engine.api.process.ProcessName
 import pl.touk.nussknacker.engine.management.FlinkConfig
 import pl.touk.nussknacker.engine.management.periodic.flink.FlinkJarManager
+import pl.touk.nussknacker.engine.management.periodic.service.{AdditionalDeploymentDataProvider, PeriodicProcessListener}
 import slick.jdbc
 import slick.jdbc.JdbcProfile
 import sttp.client.asynchttpclient.future.AsyncHttpClientFutureBackend
@@ -27,7 +28,9 @@ object PeriodicProcessManager {
             periodicBatchConfig: PeriodicBatchConfig,
             flinkConfig: FlinkConfig,
             originalConfig: Config,
-            modelData: ModelData): PeriodicProcessManager = {
+            modelData: ModelData,
+            listener: PeriodicProcessListener,
+            additionalDeploymentDataProvider: AdditionalDeploymentDataProvider): PeriodicProcessManager = {
     implicit val system: ActorSystem = ActorSystem("periodic-process-manager-provider")
     implicit val ec: ExecutionContext = ExecutionContext.global
     implicit val backend: SttpBackend[Future, Nothing, NothingT] = AsyncHttpClientFutureBackend.usingConfigBuilder { builder =>
@@ -37,7 +40,7 @@ object PeriodicProcessManager {
     val (db: jdbc.JdbcBackend.DatabaseDef, dbProfile: JdbcProfile) = DbInitializer.init(periodicBatchConfig.db)
     val scheduledProcessesRepository = new SlickPeriodicProcessesRepository(db, dbProfile)
     val jarManager = FlinkJarManager(flinkConfig, periodicBatchConfig, modelData, enrichDeploymentWithJarDataFactory(originalConfig))
-    val service = new PeriodicProcessService(delegate, jarManager, scheduledProcessesRepository)
+    val service = new PeriodicProcessService(delegate, jarManager, scheduledProcessesRepository, listener, additionalDeploymentDataProvider)
     system.actorOf(DeploymentActor.props(service, periodicBatchConfig.deployInterval))
     system.actorOf(RescheduleFinishedActor.props(service, periodicBatchConfig.rescheduleCheckInterval))
     val toClose = () => {
@@ -108,8 +111,8 @@ class PeriodicProcessManager(delegate: ProcessManager,
   override def findJobStatus(name: ProcessName): Future[Option[ProcessState]] = {
     def handleScheduled(original: Option[ProcessState]): Future[Option[ProcessState]] = {
       service.getScheduledRunDetails(name).map { maybeScheduledRunDetails =>
-        maybeScheduledRunDetails.map { scheduledRunDetails =>
-          scheduledRunDetails.status match {
+        maybeScheduledRunDetails.map { case (scheduledRunDetails, status) =>
+          status match {
             case PeriodicProcessDeploymentStatus.Scheduled => Some(ProcessState(
               Some(ExternalDeploymentId("future")),
               status = ScheduledStatus(scheduledRunDetails.runAt),