+++ engine/flink/management/periodic/src/main/scala/pl/touk/nussknacker/engine/management/periodic/db/PeriodicProcessesTable.scala
@@ -25,8 +25,6 @@ trait PeriodicProcessesTableFactory {
 
     def modelConfig: Rep[String] = column[String]("model_config", NotNull)
 
-    def buildInfoJson: Rep[String] = column[String]("build_info_json", NotNull)
-
     def jarFileName: Rep[String] = column[String]("jar_file_name", NotNull)
 
     def periodicProperty: Rep[String] = column[String]("periodic_property", NotNull)
@@ -35,7 +33,7 @@ trait PeriodicProcessesTableFactory {
 
     def createdAt: Rep[LocalDateTime] = column[LocalDateTime]("created_at", NotNull)
 
-    override def * : ProvenShape[PeriodicProcessEntity] = (id, processName, processVersionId, processJson, modelConfig, buildInfoJson, jarFileName, periodicProperty, active, createdAt) <>
+    override def * : ProvenShape[PeriodicProcessEntity] = (id, processName, processVersionId, processJson, modelConfig, jarFileName, periodicProperty, active, createdAt) <>
       ((PeriodicProcessEntity.apply _).tupled, PeriodicProcessEntity.unapply)
   }
 
@@ -48,7 +46,6 @@ case class PeriodicProcessEntity(id: PeriodicProcessId,
                                  processVersionId: Long,
                                  processJson: String,
                                  modelConfig: String,
-                                 buildInfoJson: String,
                                  jarFileName: String,
                                  periodicProperty: String,
                                  active: Boolean,