+++ engine/flink/management/periodic/src/test/scala/pl/touk/nussknacker/engine/management/periodic/JarManagerStub.scala
@@ -1,20 +1,21 @@
 package pl.touk.nussknacker.engine.management.periodic
 
 import pl.touk.nussknacker.engine.api.ProcessVersion
+import pl.touk.nussknacker.engine.api.deployment.{DeploymentData, ExternalDeploymentId}
 
 import scala.concurrent.Future
 
 class JarManagerStub extends JarManager {
 
-  var deployWithJarFuture: Future[Unit] = Future.successful(())
+  var deployWithJarFuture: Future[Option[ExternalDeploymentId]] = Future.successful(None)
 
   override def prepareDeploymentWithJar(processVersion: ProcessVersion, processJson: String): Future[DeploymentWithJarData] = {
     Future.successful(
-      DeploymentWithJarData(processVersion = processVersion, processJson = processJson, modelConfig = "", buildInfoJson = "{}", jarFileName = "")
+      DeploymentWithJarData(processVersion = processVersion, processJson = processJson, modelConfig = "", jarFileName = "")
     )
   }
 
-  override def deployWithJar(deploymentWithJarData: DeploymentWithJarData): Future[Unit] = deployWithJarFuture
+  override def deployWithJar(deploymentWithJarData: DeploymentWithJarData, deploymentData: DeploymentData): Future[Option[ExternalDeploymentId]] = deployWithJarFuture
 
   override def deleteJar(jarFileName: String): Future[Unit] = Future.successful(())
 }